requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
httpx==0.27.0
google-cloud-bigquery==3.17.2
google-auth==2.28.1
pandas==2.2.1
openpyxl==3.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
